Add a new BackingButton class
A BackingButton is a 'dead' button that serves only visual purposes. It mirrors the size of its parent region, and looks like a backing plate behind another node. The button is 'dead' in the sense that it cannot hold the focus of a window, it cannot self-arm, and it cannot respond to any user input.