Fix bug with differently-sized cards cross-sliding
When cards with very different sizes (more than a factor of 2) slide
over each other, the positioning and placement of the dragging card is
offset when its own height is counted against it. The primary content of
this change exists in the getDragCardIndex() method, which no longer
counts the height of the card against itself.